+++ CHANGES.md
@@ -1,6 +1,7 @@
 ﻿### 3.29.4 (2023-06-07 13:45:00 UTC)
 
 * Change allow Python 3.8.17, 3.9.17, 3.10.12
+* Fix setting airtime timezone for "Wanted" episodes during auto search
 
 
 ### 3.29.3 (2023-05-31 13:30:00 UTC)